+++ projects/batfish-common-protocol/src/main/java/org/batfish/specifier/parboiled/NodeNoMatchMessages.java
@@ -1,9 +1,9 @@
 package org.batfish.specifier.parboiled;
 
-import static org.batfish.datamodel.answers.InputValidationUtils.getErrorMessageEmptyNameRegex;
-import static org.batfish.datamodel.answers.InputValidationUtils.getErrorMessageMissingBook;
-import static org.batfish.datamodel.answers.InputValidationUtils.getErrorMessageMissingGroup;
-import static org.batfish.datamodel.answers.InputValidationUtils.getErrorMessageMissingName;
+import static org.batfish.specifier.parboiled.InputValidationUtils.getErrorMessageEmptyNameRegex;
+import static org.batfish.specifier.parboiled.InputValidationUtils.getErrorMessageMissingBook;
+import static org.batfish.specifier.parboiled.InputValidationUtils.getErrorMessageMissingGroup;
+import static org.batfish.specifier.parboiled.InputValidationUtils.getErrorMessageMissingName;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
