+++ projects/coordinator/src/main/java/org/batfish/coordinator/WorkMgr.java
@@ -98,7 +98,6 @@ import org.batfish.datamodel.answers.AnswerSummary;
 import org.batfish.datamodel.answers.AutoCompleteUtils;
 import org.batfish.datamodel.answers.AutocompleteSuggestion;
 import org.batfish.datamodel.answers.InputValidationNotes;
-import org.batfish.datamodel.answers.InputValidationUtils;
 import org.batfish.datamodel.answers.Issue;
 import org.batfish.datamodel.answers.Schema;
 import org.batfish.datamodel.answers.Schema.Type;
@@ -123,6 +122,7 @@ import org.batfish.identifiers.QuestionId;
 import org.batfish.identifiers.SnapshotId;
 import org.batfish.referencelibrary.ReferenceLibrary;
 import org.batfish.role.NodeRolesData;
+import org.batfish.specifier.parboiled.InputValidationUtils;
 import org.batfish.storage.StorageProvider;
 import org.batfish.storage.StoredObjectMetadata;
 import org.codehaus.jettison.json.JSONException;
